feat(steps): support w3c WAI-ARIA pattern
