+++ packages/docs/components/Dropdown.md
@@ -47,7 +47,7 @@ The component implements the W3C ARIA APG [Combobox Pattern](https://www.w3.org/
 | ariaLabel        | Accessibility aria-label to be passed to the trigger element - usefull if selectable                                                                                             | string                                                                                                           | -                                                                                                | <code style='white-space: nowrap; padding: 0;'></code>                                                                                                         |
 | checkScroll      | Makes the component check if menu reached scroll start or end and emit scroll events.                                                                                            | boolean                                                                                                          | -                                                                                                | <div><small>From <b>config</b>:</small></div><code style='white-space: nowrap; padding: 0;'>dropdown: {<br>&nbsp;&nbsp;checkScroll: false<br>}</code>          |
 | closeOnOutside   | Close Dropdown when clicked outside                                                                                                                                              | boolean                                                                                                          | -                                                                                                | <div><small>From <b>config</b>:</small></div><code style='white-space: nowrap; padding: 0;'>dropdown: {<br>&nbsp;&nbsp;closeOnOutside: true<br>}</code>        |
-| closeOnScroll    | Close Dropdown when page get scrolled                                                                                                                                            | boolean                                                                                                          | -                                                                                                | <div><small>From <b>config</b>:</small></div><code style='white-space: nowrap; padding: 0;'>dropdown: {<br>&nbsp;&nbsp;closeOnScroll: true<br>}</code>         |
+| closeOnScroll    | Close Dropdown when page get scrolled                                                                                                                                            | boolean                                                                                                          | -                                                                                                | <div><small>From <b>config</b>:</small></div><code style='white-space: nowrap; padding: 0;'>dropdown: {<br>&nbsp;&nbsp;closeOnScroll: false<br>}</code>        |
 | delay            | Dropdown delay before it appears (number in ms)                                                                                                                                  | number                                                                                                           | -                                                                                                |                                                                                                                                                                |
 | desktopModal     | Dropdown content (items) are shown into a modal on desktop                                                                                                                       | boolean                                                                                                          | -                                                                                                | <div><small>From <b>config</b>:</small></div><code style='white-space: nowrap; padding: 0;'>dropdown: {<br>&nbsp;&nbsp;desktopModal: false<br>}</code>         |
 | disabled         | Dropdown is disabled                                                                                                                                                             | boolean                                                                                                          | -                                                                                                | <code style='white-space: nowrap; padding: 0;'>false</code>                                                                                                    |
